CRA-505, CRA-559, CRA-561, CRA-568, CRA-593, CRA-668: Store large grids as float instead of double.

+++ src/analyzelog.h
@@ -20,21 +20,21 @@ class Analyzelog {
 public:
 
 
-  Analyzelog(const std::vector<NRLib::Well>                           & wells,
-             const std::map<std::string, BlockedLogsCommon *>         & mapped_blocked_logs_for_correlation,
-             const std::vector<NRLib::Grid<double> *>                 & background,
-             //const std::vector<NRLib::Grid<double> >                  & background_max_Hz,
-             const Simbox                                             * simbox,
-             const ModelSettings                                      * model_settings,
-             std::string                                              & err_txt);
+  Analyzelog(const std::vector<NRLib::Well>                          & wells,
+             const std::map<std::string, BlockedLogsCommon *>        & mapped_blocked_logs_for_correlation,
+             const std::vector<NRLib::Grid<float> *>                 & background,
+             //const std::vector<NRLib::Grid<double> >                 & background_max_Hz,
+             const Simbox                                            * simbox,
+             const ModelSettings                                     * model_settings,
+             std::string                                             & err_txt);
 
-  Analyzelog(const std::vector<NRLib::Well>                           & wells,
-             const std::map<std::string, BlockedLogsCommon *>         & mapped_blocked_logs_for_correlation,
-             const std::vector<std::vector<NRLib::Grid<double> *> >   & background,
-             //const std::vector<std::vector<NRLib::Grid<double> > >    & background_max_Hz,
-             const std::vector<Simbox>                                & interval_simboxes,
-             const ModelSettings                                      * model_settings,
-             std::string                                              & err_txt);
+  Analyzelog(const std::vector<NRLib::Well>                          & wells,
+             const std::map<std::string, BlockedLogsCommon *>        & mapped_blocked_logs_for_correlation,
+             const std::vector<std::vector<NRLib::Grid<float> *> >   & background,
+             //const std::vector<std::vector<NRLib::Grid<double> > >   & background_max_Hz,
+             const std::vector<Simbox>                               & interval_simboxes,
+             const ModelSettings                                     * model_settings,
+             std::string                                             & err_txt);
 
   /*
   Analyzelog(std::vector<WellData *>  & wells,
@@ -55,13 +55,13 @@ public:
 
 private:
 
-  void            EstimateLnData(std::vector<std::vector<float> >                      & log_data,
-                                 const NRLib::Grid<double>                             * background,
-                                 const std::vector<std::string>                        & well_names,
-                                 const std::map<std::string, BlockedLogsCommon *>      & mapped_blocked_logs_for_correlation,
-                                 const std::string                                     & interval_name,
-                                 const std::string                                     & log_name,
-                                 std::string                                           & err_txt);
+  void            EstimateLnData(std::vector<std::vector<float> >                     & log_data,
+                                 const NRLib::Grid<float>                             * background,
+                                 const std::vector<std::string>                       & well_names,
+                                 const std::map<std::string, BlockedLogsCommon *>     & mapped_blocked_logs_for_correlation,
+                                 const std::string                                    & interval_name,
+                                 const std::string                                    & log_name,
+                                 std::string                                          & err_txt);
 
   void            EstimateCorrelation(const ModelSettings                                      * model_settings,
                                       const std::vector<NRLib::Well>                           & wells,
@@ -70,7 +70,7 @@ private:
                                       std::string                                              & interval_name,
                                       const Simbox                                             * interval_simbox,
                                       bool                                                     & enough_data_for_corr_estimation,
-                                      const std::vector<NRLib::Grid<double> *>                  & background,
+                                      const std::vector<NRLib::Grid<float> *>                  & background,
                                       std::string                                              & errTxt);
 
     void          EstimateCorrelation(const ModelSettings                                      * model_settings,
